Refactor api into its own file (#6)
* move api classes and support into their own file

* fix analytics to properly import API
